feat: Add support for customized tailwind cli asset name
The code changes introduce a new setting, `TAILWIND_CLI_ASSET_NAME`. The last release introduced `TAILWIND_CLI_SRC_REPO`, but that alone is not enough to pull from a customized repository. The repository might be using a custom name for the Tailwind CLI file, such as `tailwind-extra`. This case was not taken into account, an oversight on my part.